Task scheduler: revenue calculator

+++ src/main/java/ro/msg/learning/shop/dto/ProductDto.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -14,7 +12,7 @@ public class ProductDto {
     private int id;
     private String name;
     private String description;
-    private BigDecimal price;
+    private double price;
     private double weight;
     private String categoryName;
     private String supplierName;
